refactor(*): set up package structure according to mvvm pattern
