+++ app/src/main/java/com/csapp/sqli/repository/DatabaseRepository.kt
@@ -1,4 +1,4 @@
-package com.csapp.sqli
+package com.csapp.sqli.repository
 
 import android.annotation.SuppressLint
 import android.content.Context
@@ -6,7 +6,7 @@ import android.database.sqlite.SQLiteDatabase
 import android.database.sqlite.SQLiteOpenHelper
 import android.util.Log
 
-class DatabaseHelper(context: Context?) :
+class DatabaseRepository(context: Context?) :
     SQLiteOpenHelper(context, DATABASE_NAME, null, DATABASE_VERSION) {
     override fun onCreate(p0: SQLiteDatabase?) {
         // Not yet implemented
